adjust inner calls recycler view data and merging conditions

+++ app/src/main/java/com/example/whatsappclonelombenis/CallInfoRecViewAdapter.java
@@ -53,7 +53,7 @@ public class CallInfoRecViewAdapter extends RecyclerView.Adapter<CallInfoRecView
         holder.callTime.setText(callTimeText);
 
         holder.callDuration.setText(Integer.toString(calls.get(position).getCallDuration()));
-        holder.callMB.setText(Integer.toString(calls.get(position).getCallMB()));
+        holder.callMB.setText(Double.toString(calls.get(position).getCallMB()));
     }
 
     @Override
